+++ src/main/java/net/lenni0451/rivet/theme/Theme.java
@@ -21,8 +21,8 @@ public abstract class Theme {
     public static final ThemeKey<Color> TEXT_COLOR = register("general.text_color", Color.class);
 
     // Button
-    public static final ThemeKey<Integer> BUTTON_CORNER_RADIUS = register("button.corner_radius", Integer.class);
-    public static final ThemeKey<Integer> BUTTON_OUTLINE_WIDTH = register("button.outline_width", Integer.class);
+    public static final ThemeKey<Float> BUTTON_CORNER_RADIUS = register("button.corner_radius", Float.class);
+    public static final ThemeKey<Float> BUTTON_OUTLINE_WIDTH = register("button.outline_width", Float.class);
     public static final ThemeKey<Color> BUTTON_INACTIVE_COLOR = register("button.inactive_color", Color.class);
     public static final ThemeKey<Color> BUTTON_INACTIVE_OUTLINE_COLOR = register("button.inactive_outline_color", Color.class);
     public static final ThemeKey<Color> BUTTON_ACTIVE_COLOR = register("button.active_color", Color.class);
